[Enhancement] Option to have different Pdf templates for other PageTypes
Based on the Object ClassName: you can now have differing pdf print templates for pagetypes, with a fallback to the default template (PdfPage.ss)

For example, of you have extended Page.php for your HomePage (ClassName: 'HomePage'), you can have 'HomePage_PdfPage.ss' as an extra printlayout. If this file is not present, it will falback to 'PdfPage.ss'.